Fix kerenel cmdline parsing
Also i found that we do the parsing in 2 differnet places
I refactored the code for reusability
